map is an item of the recyclerview
