+++ app/src/main/res/layout/my_flags_layout.xml
@@ -45,19 +45,19 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical">
-
+<!--
         <FrameLayout
             android:id="@+id/my_map_holder"
             android:layout_width="match_parent"
             android:layout_height="0dp"
             android:layout_weight="0.4"
             />
-
+-->
         <com.gcw.sapienza.places.layouts.MSwipeRefreshLayout xmlns:tools="http://schemas.android.com/tools"
             android:id="@+id/my_swipe_refresh"
             android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="0.6"
+            android:layout_height="match_parent"
+            android:layout_weight="1.0"
             android:orientation="vertical"
             tools:context="com.gcw.sapienza.places.activities.MainActivity" />
